# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AiStatus {
  provider: string;
  model: string;
  configured: boolean;
  client: string;
}

/**
 * @nullable
 */
export type PlanningRequestPlanningStyle = typeof PlanningRequestPlanningStyle[keyof typeof PlanningRequestPlanningStyle] | null;


export const PlanningRequestPlanningStyle = {
  lighter: 'lighter',
  balanced: 'balanced',
  fuller: 'fuller',
} as const;

export interface PlanningTimeWindow {
  /** @pattern ^\d{2}:\d{2}$ */
  startTime: string;
  /** @pattern ^\d{2}:\d{2}$ */
  endTime: string;
}

export interface PlanningRequest {
  /**
     * @minLength 1
     * @maxLength 2000
     */
  intention: string;
  /** @pattern ^\d{4}-\d{2}-\d{2}$ */
  currentDate: string;
  /** @pattern ^\d{2}:\d{2}$ */
  currentTime: string;
  /** @maxItems 12 */
  availableTime: PlanningTimeWindow[];
  /** @nullable */
  planningStyle?: PlanningRequestPlanningStyle;
  /**
     * @maxLength 2000
     * @nullable
     */
  fixedCommitments?: string | null;
  useHistoricalContext?: boolean;
  /**
     * @maxLength 4000
     * @nullable
     */
  historicalContext?: string | null;
}

/**
 * @nullable
 */
export type ProposedActivityCategory = typeof ProposedActivityCategory[keyof typeof ProposedActivityCategory] | null;


export const ProposedActivityCategory = {
  work: 'work',
  recovery: 'recovery',
  managing: 'managing',
  social: 'social',
  fun: 'fun',
} as const;

export interface ProposedActivity {
  /**
     * @minLength 1
     * @maxLength 200
     */
  title: string;
  /** @pattern ^\d{4}-\d{2}-\d{2}$ */
  scheduledDate: string;
  /** @pattern ^\d{2}:\d{2}$ */
  startTime: string;
  /** @pattern ^\d{2}:\d{2}$ */
  endTime: string;
  /** @nullable */
  category: ProposedActivityCategory;
  /**
     * @maxLength 500
     * @nullable
     */
  note: string | null;
}

export interface PlanningTimeBlock {
  /** @pattern ^\d{4}-\d{2}-\d{2}$ */
  scheduledDate: string;
  /** @pattern ^\d{2}:\d{2}$ */
  startTime: string;
  /** @pattern ^\d{2}:\d{2}$ */
  endTime: string;
  /**
     * @minLength 1
     * @maxLength 500
     */
  reason: string;
}

export interface PlanningConflict {
  /**
     * @minLength 1
     * @maxLength 500
     */
  description: string;
  /**
     * @maxItems 10
     * @items.minLength 1
     * @items.maxLength 200
     */
  relatedActivityTitles: string[];
}

export interface PlanningNotFittedItem {
  /**
     * @minLength 1
     * @maxLength 200
     */
  title: string;
  /**
     * @minLength 1
     * @maxLength 500
     */
  reason: string;
}

export interface PlanningProposal {
  /** @maxItems 50 */
  proposedActivities: ProposedActivity[];
  /** @maxItems 50 */
  buffers: PlanningTimeBlock[];
  /** @maxItems 50 */
  restPeriods: PlanningTimeBlock[];
  /**
     * @maxItems 20
     * @items.minLength 1
     * @items.maxLength 500
     */
  assumptions: string[];
  /** @maxItems 20 */
  conflicts: PlanningConflict[];
  /** @maxItems 50 */
  didNotFit: PlanningNotFittedItem[];
}

/**
 * @nullable
 */
export type ActivityCategory = typeof ActivityCategory[keyof typeof ActivityCategory] | null;


export const ActivityCategory = {
  work: 'work',
  recovery: 'recovery',
  managing: 'managing',
  social: 'social',
  fun: 'fun',
} as const;

export interface Activity {
  id: number;
  title: string;
  /** @pattern ^\d{4}-\d{2}-\d{2}$ */
  scheduledDate: string;
  startTime: string;
  /** @nullable */
  endTime: string | null;
  /** @nullable */
  category: ActivityCategory;
  completed: boolean;
  locked: boolean;
  pinned: boolean;
  /** @nullable */
  note: string | null;
}

/**
 * @nullable
 */
export type ActivityInputCategory = typeof ActivityInputCategory[keyof typeof ActivityInputCategory] | null;


export const ActivityInputCategory = {
  work: 'work',
  recovery: 'recovery',
  managing: 'managing',
  social: 'social',
  fun: 'fun',
} as const;

export interface ActivityInput {
  /** @minLength 1 */
  title: string;
  /** @pattern ^\d{4}-\d{2}-\d{2}$ */
  scheduledDate: string;
  startTime: string;
  /** @nullable */
  endTime?: string | null;
  /** @nullable */
  category?: ActivityInputCategory;
  completed?: boolean;
  locked?: boolean;
  pinned?: boolean;
  /** @nullable */
  note?: string | null;
}

/**
 * @nullable
 */
export type ActivityUpdateCategory = typeof ActivityUpdateCategory[keyof typeof ActivityUpdateCategory] | null;


export const ActivityUpdateCategory = {
  work: 'work',
  recovery: 'recovery',
  managing: 'managing',
  social: 'social',
  fun: 'fun',
} as const;

export interface ActivityUpdate {
  /** @minLength 1 */
  title?: string;
  /** @pattern ^\d{4}-\d{2}-\d{2}$ */
  scheduledDate?: string;
  startTime?: string;
  /** @nullable */
  endTime?: string | null;
  /** @nullable */
  category?: ActivityUpdateCategory;
  completed?: boolean;
  locked?: boolean;
  pinned?: boolean;
  /** @nullable */
  note?: string | null;
}

export type ActivityChangeChangeType = typeof ActivityChangeChangeType[keyof typeof ActivityChangeChangeType];


export const ActivityChangeChangeType = {
  moved: 'moved',
  extended: 'extended',
  shortened: 'shortened',
  renamed: 'renamed',
  removed: 'removed',
  replaced: 'replaced',
  completed_later: 'completed_later',
  review_note: 'review_note',
} as const;

export type ActivityChangeSource = typeof ActivityChangeSource[keyof typeof ActivityChangeSource];


export const ActivityChangeSource = {
  manual: 'manual',
  ai_approved: 'ai_approved',
} as const;

export interface ActivityChange {
  id: number;
  /** @nullable */
  activityId: number | null;
  /** @pattern ^\d{4}-\d{2}-\d{2}$ */
  scheduledDate: string;
  /** @nullable */
  activityTitle: string | null;
  changeType: ActivityChangeChangeType;
  /** @nullable */
  previousTitle: string | null;
  /** @nullable */
  nextTitle: string | null;
  /** @nullable */
  previousStartTime: string | null;
  /** @nullable */
  nextStartTime: string | null;
  /** @nullable */
  previousEndTime: string | null;
  /** @nullable */
  nextEndTime: string | null;
  /** @nullable */
  note: string | null;
  source: ActivityChangeSource;
  changedAt: string;
}

export interface ActivityChangeNoteInput {
  /** @pattern ^\d{4}-\d{2}-\d{2}$ */
  scheduledDate: string;
  /**
     * @minLength 1
     * @maxLength 1000
     */
  note: string;
}

export type ListActivitiesParams = {
/**
 * @pattern ^\d{4}-\d{2}-\d{2}$
 */
date: string;
};

export type ListActivityChangesParams = {
/**
 * @pattern ^\d{4}-\d{2}-\d{2}$
 */
date: string;
};