# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: ai
    description: Server-side AI configuration operations
  - name: planning
    description: Bounded AI planning proposal operations
  - name: activities
    description: Day Tripper activity operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /ai/status:
    get:
      operationId: getAiStatus
      tags: [ai]
      summary: Get AI configuration status
      description: Returns safe metadata about the server-side Gemini configuration without exposing credentials
      responses:
        "200":
          description: AI configuration status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiStatus"
  /planning/proposals:
    post:
      operationId: createPlanningProposal
      tags: [planning]
      summary: Create a bounded planning proposal
      description: Creates a validated AI proposal without changing saved activities
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlanningRequest"
      responses:
        "200":
          description: Validated planning proposal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlanningProposal"
        "400":
          description: Invalid planning request
        "502":
          description: The planning provider returned an invalid proposal
        "503":
          description: Planning service is unavailable
  /activities:
    get:
      operationId: listActivities
      tags: [activities]
      summary: List activities for a day
      parameters:
        - name: date
          in: query
          required: true
          schema:
            type: string
            pattern: '^\d{4}-\d{2}-\d{2}$'
      responses:
        "200":
          description: Activities for the requested day
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Activity"
    post:
      operationId: createActivity
      tags: [activities]
      summary: Add an activity to a day
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActivityInput"
      responses:
        "201":
          description: Activity created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Activity"
  /activities/{id}:
    patch:
      operationId: updateActivity
      tags: [activities]
      summary: Update an activity
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActivityUpdate"
      responses:
        "200":
          description: Activity updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Activity"
        "404":
          description: Activity not found
    delete:
      operationId: deleteActivity
      tags: [activities]
      summary: Delete an activity
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Activity deleted
        "404":
          description: Activity not found
  /activity-changes:
    get:
      operationId: listActivityChanges
      tags: [activities]
      summary: List manual changes for a day
      description: Returns a private, chronological record of schedule changes for the requested day
      parameters:
        - name: date
          in: query
          required: true
          schema:
            type: string
            pattern: '^\d{4}-\d{2}-\d{2}$'
      responses:
        "200":
          description: Schedule changes for the requested day
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityChange"
    post:
      operationId: addActivityChangeNote
      tags: [activities]
      summary: Add an optional note to the change review
      description: Records an optional user explanation without changing the saved schedule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActivityChangeNoteInput"
      responses:
        "201":
          description: Review note saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActivityChange"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    AiStatus:
      type: object
      properties:
        provider:
          type: string
        model:
          type: string
        configured:
          type: boolean
        client:
          type: string
      required:
        - provider
        - model
        - configured
        - client
    PlanningRequest:
      type: object
      properties:
        intention:
          type: string
          minLength: 1
          maxLength: 2000
        currentDate:
          type: string
          pattern: '^\d{4}-\d{2}-\d{2}$'
        currentTime:
          type: string
          pattern: '^\d{2}:\d{2}$'
        availableTime:
          type: array
          maxItems: 12
          items:
            $ref: "#/components/schemas/PlanningTimeWindow"
        planningStyle:
          type: ["string", "null"]
          enum: [lighter, balanced, fuller, null]
        fixedCommitments:
          type: ["string", "null"]
          maxLength: 2000
        useHistoricalContext:
          type: boolean
          default: false
        historicalContext:
          type: ["string", "null"]
          maxLength: 4000
      required:
        - intention
        - currentDate
        - currentTime
        - availableTime
    PlanningTimeWindow:
      type: object
      properties:
        startTime:
          type: string
          pattern: '^\d{2}:\d{2}$'
        endTime:
          type: string
          pattern: '^\d{2}:\d{2}$'
      required:
        - startTime
        - endTime
    PlanningProposal:
      type: object
      properties:
        proposedActivities:
          type: array
          maxItems: 50
          items:
            $ref: "#/components/schemas/ProposedActivity"
        buffers:
          type: array
          maxItems: 50
          items:
            $ref: "#/components/schemas/PlanningTimeBlock"
        restPeriods:
          type: array
          maxItems: 50
          items:
            $ref: "#/components/schemas/PlanningTimeBlock"
        assumptions:
          type: array
          maxItems: 20
          items:
            type: string
            minLength: 1
            maxLength: 500
        conflicts:
          type: array
          maxItems: 20
          items:
            $ref: "#/components/schemas/PlanningConflict"
        didNotFit:
          type: array
          maxItems: 50
          items:
            $ref: "#/components/schemas/PlanningNotFittedItem"
      required:
        - proposedActivities
        - buffers
        - restPeriods
        - assumptions
        - conflicts
        - didNotFit
    ProposedActivity:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
        scheduledDate:
          type: string
          pattern: '^\d{4}-\d{2}-\d{2}$'
        startTime:
          type: string
          pattern: '^\d{2}:\d{2}$'
        endTime:
          type: string
          pattern: '^\d{2}:\d{2}$'
        category:
          type: ["string", "null"]
          enum: [work, recovery, managing, social, fun, null]
        note:
          type: ["string", "null"]
          maxLength: 500
      required:
        - title
        - scheduledDate
        - startTime
        - endTime
        - category
        - note
    PlanningTimeBlock:
      type: object
      properties:
        scheduledDate:
          type: string
          pattern: '^\d{4}-\d{2}-\d{2}$'
        startTime:
          type: string
          pattern: '^\d{2}:\d{2}$'
        endTime:
          type: string
          pattern: '^\d{2}:\d{2}$'
        reason:
          type: string
          minLength: 1
          maxLength: 500
      required:
        - scheduledDate
        - startTime
        - endTime
        - reason
    PlanningConflict:
      type: object
      properties:
        description:
          type: string
          minLength: 1
          maxLength: 500
        relatedActivityTitles:
          type: array
          maxItems: 10
          items:
            type: string
            minLength: 1
            maxLength: 200
      required:
        - description
        - relatedActivityTitles
    PlanningNotFittedItem:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
        reason:
          type: string
          minLength: 1
          maxLength: 500
      required:
        - title
        - reason
    Activity:
      type: object
      properties:
        id:
          type: number
        title:
          type: string
        scheduledDate:
          type: string
          pattern: '^\d{4}-\d{2}-\d{2}$'
        startTime:
          type: string
        endTime:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
          enum: [work, recovery, managing, social, fun, null]
        completed:
          type: boolean
        locked:
          type: boolean
        pinned:
          type: boolean
        note:
          type: ["string", "null"]
      required:
        - id
        - title
        - scheduledDate
        - startTime
        - endTime
        - category
        - completed
        - locked
        - pinned
        - note
    ActivityInput:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        scheduledDate:
          type: string
          pattern: '^\d{4}-\d{2}-\d{2}$'
        startTime:
          type: string
        endTime:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
          enum: [work, recovery, managing, social, fun, null]
        completed:
          type: boolean
        locked:
          type: boolean
        pinned:
          type: boolean
        note:
          type: ["string", "null"]
      required:
        - title
        - scheduledDate
        - startTime
    ActivityUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        scheduledDate:
          type: string
          pattern: '^\d{4}-\d{2}-\d{2}$'
        startTime:
          type: string
        endTime:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
          enum: [work, recovery, managing, social, fun, null]
        completed:
          type: boolean
        locked:
          type: boolean
        pinned:
          type: boolean
        note:
          type: ["string", "null"]
    ActivityChange:
      type: object
      properties:
        id:
          type: number
        activityId:
          type: ["number", "null"]
        scheduledDate:
          type: string
          pattern: '^\d{4}-\d{2}-\d{2}$'
        activityTitle:
          type: ["string", "null"]
        changeType:
          type: string
          enum: [moved, extended, shortened, renamed, removed, replaced, completed_later, review_note]
        previousTitle:
          type: ["string", "null"]
        nextTitle:
          type: ["string", "null"]
        previousStartTime:
          type: ["string", "null"]
        nextStartTime:
          type: ["string", "null"]
        previousEndTime:
          type: ["string", "null"]
        nextEndTime:
          type: ["string", "null"]
        note:
          type: ["string", "null"]
        source:
          type: string
          enum: [manual, ai_approved]
        changedAt:
          type: string
      required:
        - id
        - activityId
        - scheduledDate
        - activityTitle
        - changeType
        - previousTitle
        - nextTitle
        - previousStartTime
        - nextStartTime
        - previousEndTime
        - nextEndTime
        - note
        - source
        - changedAt
    ActivityChangeNoteInput:
      type: object
      properties:
        scheduledDate:
          type: string
          pattern: '^\d{4}-\d{2}-\d{2}$'
        note:
          type: string
          minLength: 1
          maxLength: 1000
      required:
        - scheduledDate
        - note